refactor(ledger): collocate the organization route registrars :hammer:
Both organization registrars already sat in organization_handler.go; they move
together into organization_routes.go so the handler file keeps only the handler
struct, the request/response envelopes and the cores, matching every resource
collocated so far.

organization_handler.go no longer needs the lib-auth middleware or fiber
imports once the guard-chain registrars leave.

No route, authz tuple, operation ID or path changed.
